Add split methods with condition to SplitUtils, version bump
Added two public methods, `splitToArray` and `splitToList`, and a private method, `splitToStream` to `SplitUtils` class. These methods allow the splitting of a string based on a given condition. This provides more flexible splitting options for users.

Also increased the version of the project in `pom.xml` from `1.0.10` to `1.0.11` to reflect the added features and improvements.