Add BooleanUtils for null-friendly Boolean checks
This commit adds a new utility class, BooleanUtils, that extends commons-lang3's BooleanUtils. It introduces two methods, isTrueOrNull and isFalseOrNull, which check if a Boolean value is true or null, or false or null, respectively. This extension is helpful when dealing with Boolean values that may be null- providing a null-friendly way to handle these cases. The README has also been updated to reflect this addition. Furthermore, the version in pom.xml has been updated from 1.0.11 to 1.0.12.